feat(scripts): Create scripts package
Creating a new package named `@spinnaker/scripts` for grouping all build-related tooling.
This package takes care of providing rollup config, relevant dependencies and exposes
scripts that could be utilized directly from deck plugins and modules via npm scripts.

The suggested approach is to add `@spinnaker/pluginsdk` as a dependency to the spinnaker
plugin or `@spinnaker/scripts` to a spinnaker module and add the following to
`package.json`.

```
{
  ...
  "scripts": {
    "start": "spinnaker-scripts start",
    "build": "spinnaker-scripts build"
  }
}
```

By default `spinnaker-scripts` will look for `rollup.config.js` file in the project directory.
If none is found there, then the default rollup config for plugins (defined in
`/config/basePluginRollupConfig.js`) will be used. Alternatively, the user can provide
the rollup config file via `-f` option.

Plugins can re-use `@spinnaker/pluginsdk/pluginconfig/rollup.confg.js` and customize their
bundling. Modules can do that with `@spinnaker/scripts/config/baseModuleRollupConfig.js`.
