+++ packages/pluginsdk/package.json
@@ -7,8 +7,8 @@
   "license": "Apache-2.0",
   "scripts": {
     "clean": "npx shx rm -rf dist",
-    "build": "rollup -c",
-    "watch": "rollup -c -w",
+    "build": "spinnaker-scripts build",
+    "watch": "spinnaker-scripts start",
     "prepublishOnly": "npm run build"
   },
   "bin": {
@@ -20,6 +20,7 @@
     "write-json": "scripts/write-json.js"
   },
   "dependencies": {
+    "@spinnaker/scripts": "^0.0.1",
     "check-peer-dependencies": "^4.0.0",
     "chokidar": "^3.5.1",
     "express": "^4.17.1",
@@ -33,15 +34,8 @@
     "yargs": "^16.2.0"
   },
   "devDependencies": {
-    "@rollup/plugin-commonjs": "17.1.0",
-    "@rollup/plugin-node-resolve": "11.1.1",
-    "@rollup/plugin-typescript": "8.1.1",
     "@spinnaker/core": "0.0.549",
     "@types/yargs": "^16.0.0",
-    "rollup": "2.39.0",
-    "rollup-plugin-external-globals": "0.6.1",
-    "rollup-plugin-less": "1.1.3",
-    "rollup-plugin-postcss": "4.0.0",
     "typescript": "4.1.5"
   },
   "files": [
