+++ packages/pluginsdk/pluginconfig/rollup.config.js
@@ -1,0 +1,1 @@
+module.exports = require('@spinnaker/scripts/config/rollup.config.base.plugin');
