+++ packages/pluginsdk/rollup.config.js
@@ -1,12 +1,8 @@
-import nodeResolve from '@rollup/plugin-node-resolve';
-import commonjs from '@rollup/plugin-commonjs';
-import typescript from '@rollup/plugin-typescript';
+const baseRollupConfig = require('@spinnaker/scripts/config/rollup.config.base.module');
 
-export default [
-  {
-    input: 'src/index.ts',
-    plugins: [nodeResolve(), commonjs(), typescript()],
-    external: ['@spinnaker/core'],
-    output: [{ dir: 'dist', format: 'es', sourcemap: true }],
-  },
-];
+module.exports = {
+  ...baseRollupConfig,
+  external: ['@spinnaker/core'],
+  input: 'src/index.ts',
+  output: [{ dir: 'dist', format: 'es', sourcemap: true }],
+};
