+++ packages/scripts/package.json
@@ -23,6 +23,7 @@
     "ora": "^5.4.0",
     "postcss-nested": "^4.2.1",
     "rollup": "^2.35.1",
+    "rollup-plugin-auto-external": "^2.0.0",
     "rollup-plugin-external-globals": "0.6.1",
     "rollup-plugin-less": "1.1.3",
     "rollup-plugin-postcss": "3.1.8",
